Move parseRetryAfter to RetryUtil class
